backend/requirements.txt
pytest==7.1.3
SQLAlchemy==1.4.36
protobuf==3.20.1
farm-haystack==1.9.0
python-dotenv==0.21.0
psycopg2==2.9.3
more-itertools==8.13.0
spacy==3.4.1
setuptools~=50.3.1
requests~=2.24.0
iso-639==0.4.5

# backend
uvicorn~=0.18.3
fastapi~=0.85.1
pydantic~=1.9.2
starlette~=0.20.4
fastapi-cache2[redis]==0.1.9
python-multipart==0.0.5

torch==1.12 # keep torch at the bottom of the list to allow canceling in case of cuda mismatch
